refactor: split remaining workspace add families
