+++ .sampo/changesets/960-split-add-workspace-families.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split remaining workspace add variation, style, transform, and hooked-block scaffold flows into focused modules.
